feat: refactor TagsDisplay + TagsList into reuseable component, use TagsDisplay on PostList + Post
